# Conflicts:
#	android/demo/build.gradle
#	android/gradle/wrapper/gradle-wrapper.properties
#	android/tangram/build.gradle

Also changed #include "platform.h" to #include "log.h" where appropriate.

+++ core/src/tile/tileTask.h
@@ -86,7 +86,7 @@ protected:
 
     bool m_canceled = false;
 
-    std::atomic<double> m_priority;
+    std::atomic<float> m_priority;
     bool m_proxyState = false;
 };
